requirements.txt
Django==4.2.7
djangorestframework==3.14.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.13.0
requests==2.31.0
reportlab==4.0.7
weasyprint==60.1
python-dotenv==1.0.0
joblib==1.3.2
Pillow==10.1.0